feat: implement notification system with enhanced fetching, marking as read, and UI components
