+++ src/app/api/notifications/route.ts
@@ -1,29 +1,21 @@
 import { NextResponse } from "next/server";
 import { auth } from "@/lib/auth";
 import { db } from "@/lib/db";
+import { getNotificationsForUser } from "@/lib/notifications";
 
+// API to fetch notifications for the authenticated user
 export async function GET() {
   const session = await auth();
   if (!session?.user) {
     return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
   }
 
-  const [items, unreadCount] = await Promise.all([
-    db.notification.findMany({
-      where: { userId: session.user.id },
-      include: {
-        miniApp: { select: { id: true, slug: true, name: true } },
-      },
-      orderBy: { createdAt: "desc" },
-    }),
-    db.notification.count({
-      where: { userId: session.user.id, readAt: null },
-    }),
-  ]);
+  const payload = await getNotificationsForUser(session.user.id);
 
-  return NextResponse.json({ items, unreadCount });
+  return NextResponse.json(payload);
 }
 
+// API mark all notifications as read for the authenticated user
 export async function PATCH() {
   const session = await auth();
   if (!session?.user) {
