+++ src/app/api/notifications/unread-count/route.ts
@@ -1,7 +1,7 @@
 import { NextResponse } from "next/server";
 import { auth } from "@/lib/auth";
 import { db } from "@/lib/db";
-
+//API get count of unread notifications for the authenticated user
 export async function GET() {
   const session = await auth();
   if (!session?.user) {
